refactor(git): split git.rs monolith into module tree
Split the 3109-line git.rs file into a cmd/git/ directory with 8 files:
- mod.rs: dispatcher, shared helpers, print_help
- status.rs: run_status, parse_status, status helpers
- log.rs: run_log, parse_log
- diff/mod.rs: run_diff, DiffMode, constants
- diff/types.rs: DiffHunk, FileDiff, FileMetadata, ChangedNodeRange, etc.
- diff/parse.rs: parse_unified_diff and decomposed helpers
- diff/render.rs: AST-aware and raw hunk rendering
- diff/source.rs: file source resolution (disk, git show)
- diff/ast.rs: changed line detection, container node logic

Pure mechanical file move — zero logic changes. All 972+ tests pass,
clippy clean. include_str! paths updated for new directory depth.
